storage: move `convert_bounds` into `util::`
